Added ability to isolate <b> tags from site.
This commit had modifications allowing for the isolation of all <b></b>
tags in the site’s HTML, before writing this to local files on the
system. This is useful because all the important course information is
solely found within <b> tags. However, the source code is now far from
organized, and extra information has also tagged along. These are
issues that need to be fixed next.