Changed write_file method to write lists, not str.
The write_file() method was modified to write entire lists as opposed
to strings, as the useful find_all() method only returns lists. This
results in a much better cleaner output html file in comparison to the
previous one which was typecasted to str(). Also note how the relevant
course information only begins at ~ line 180.